Replaces game with ggez's "astroblasto"
Replaces the Rust tutorial's "guessing game" example with the ggez demo 
example "astroblasto".

This uses `ggez` version 0.5.0-rc.2, so may be unstable. It required a 
bit more work than expected: we needed to also add a dependency on 
`nalgebra` since we reference it directly, and we needed to copy *some* 
resources files from the `ggez` repo.