Splits game into library with sub-modules
Restructures the code into a library with sub-modules `ui` and
`entities`, with an entrypoint in `main.rs`.

This re-works a lot of the existing structs to untangle UI logic and
player properties, and hiding the internals of each sub-module. The
`GameState` implementation in `lib.rs` is now just defining the
entry-point and forwarding `ggez::event::EventHandler` events to
relevant sub-modules.